fix: Fix bad links in HTML reports for "Vulnerable Software & Versions"
Since the search function accepts a full CPE 2.3 name, we should be able to use whatever CPE we have, with whatever specificity the CPE has. If it happens to have a
version number in it, that will be in the URL. If it does not (due to match being against version ranges, then it will do a broad search).

Signed-off-by: Chad Wilson <29788154+chadlwilson@users.noreply.github.com>